Improved Dropbox compatibility
Implemented support for  sendMessage to let my content script talk to
the event page. Used it to improve dropbox.com/s/ page switching. It is
now improved because I discovered that the full file path was hidden in
the DOM. I can grab that instead of guessing what the path is. However,
I recently discovered that dl.dropboxusercontent.com exists and is an
awesome way for non-Dropbox viewers to see our HTML files without
having to download them or having their clicks blocked. Unfortunately
there is no hidden source path here so I have to ‘guess’ it.
Incidentally this continues to only work for emails located in
campaigns/email/discipline.