Text Search, Link Validation, and more
- Added a visual clue for link errors. Previously this only showed up
in the console.
- Implemented https://github.com/padolsey/findAndReplaceDOMText to
search the email for words should not be used depending on the context
(NS vs Sub, etc).
- Cleaned up the prep work when loading a page. All body tags are now
destroyed and replaced with a fresh one. All style blocks are now
destroyed as well instead of just the first one.
- Cleaned up the comments
- Added a new orb: Style Orb. Turn the style blocks on and off in the
dFrame and mFrame. Allows you to view what the email would look like if
an email client ignores your style block and instead only uses inline
styles.
- Dropbox share orb now just grabs the document windows URL if you’re
on the same page you’re trying to share (dl.dropboxusercontent) instead
of calling the API.
- Link validation is now pretty amazing. Lots of regex added to check
for errors in both the href as well as the CSS and images.
- Fixed Spell Check bug that would cause the squiggly lines to not
properly align itself with the corresponding text. Moving it to the
bottom of the script and adding a settimeout to the final portion of
the process did the trick.