New Orb and MailChimp Notifications
MailChimp

  - Implemented notifications for MailChimp. Now whenever you have
unfinished drafts on the campaign page, the extension will periodically
notify you. Currently set to remind you on page load of the campaign
page and again every 10 minutes regardless of whether or not you’re on
the MailChimp site. Uses chrome.storage.sync and the Event Page to
accomplish this.

Newsletter

  - New orb overlays blue PhotoShop-esque guides on the page to help
determine if your newsletter is centered, the right width, and outer
margins.
  - General enhancements to link and text checkers.