On the Webstore
With the addition of a new email team member I’ve started pushing
updates to the Chrome Webstore to share the extension. With that comes
new version numbers to handle small updates.

- Layout has been slightly updated to better fit the desktop view when
viewing Korra inside of lower resolution monitors, with extra toolbars
at the top and bottom.

- Added preliminary support for “stages”. Stages will hold
desktop/mobile views, plain text views, and more stages will be added
later like Links, Images, Accessibility, Deliverability.

- Toolbar moved to the left.

- Dropbox links leading to file downloads now redirect to the Dropbox
page instead. And HTML files viewed on Dropbox now have a link at the
top to take you to the local view of that file.

- Options page is coming along. Options are now being loaded async when
a file is viewed.

- Updates to MailChimp to make accidental sending harder.

- We’re now retrieving the HTML string straight from the original file
asynchronously instead of using the DOM rendered version that we get on
page load. We’re also using promises to detect when it’s done being
fetched.

- General bug fixes and improvements.